renamed TypeDeclarationExtractor module to TypeDeclarationExtractor
