+++ RAGESS/RAGESS/Sources/DebugView/DeclarationExtractor.swift
@@ -8,7 +8,7 @@
 import ComposableArchitecture
 import SwiftUI
 import TypeDeclaration
-import TypeDeclarationExtractor
+import DeclarationExtractor
 import XcodeObject
 
 @Reducer
